[Ref]: 플랫폼 메타데이터 서비스 리팩토링 및 구현 추가
- PlatformMetadataService 인터페이스를 정의하고, 메타데이터 조회 메서드를 선언함
- PlatformMetadataServiceImpl 클래스를 새로 추가하여 TMDB API를 통해 메타데이터를 조회하고 캐싱하는 로직을 구현함
- 예외 처리 로직을 추가하여 API 호출 시 발생할 수 있는 오류를 BusinessException으로 래핑함
